Win64 build fix

+++ appshell/appshell_extension_handler.h
@@ -74,7 +74,7 @@ CefRefPtr<CefV8Value> ListValueToV8Value(CefRefPtr<CefListValue> value, int inde
     switch (type) {
         case VTYPE_LIST: {
             CefRefPtr<CefListValue> list = value->GetList(index);
-            new_value = CefV8Value::CreateArray(list->GetSize());
+            new_value = CefV8Value::CreateArray(static_cast<int>(list->GetSize()));
             SetList(list, new_value);
         } break;
         case VTYPE_BOOL:
@@ -106,7 +106,7 @@ void SetListValue(CefRefPtr<CefV8Value> list, int index,
     switch (type) {
         case VTYPE_LIST: {
             CefRefPtr<CefListValue> listValue = value->GetList(index);
-            new_value = CefV8Value::CreateArray(listValue->GetSize());
+            new_value = CefV8Value::CreateArray(static_cast<int>(listValue->GetSize()));
             SetList(listValue, new_value);
         } break;
         case VTYPE_BOOL:
@@ -136,7 +136,7 @@ void SetListValue(CefRefPtr<CefV8Value> list, int index,
 void SetList(CefRefPtr<CefListValue> source, CefRefPtr<CefV8Value> target) {
     DCHECK(target->IsArray());
 
-    int arg_length = source->GetSize();
+    int arg_length = static_cast<int>(source->GetSize());
     if (arg_length == 0)
         return;
